Moved FileIndex class to its own file.
